+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# Version 0.7.0
+
+## Bugfixes
+
+* Add Guzzle dependency to composer.json
+* Remove wrong and invalid use statements from ContextConnection class
+
+## Features
+
+* Add PHPUnit test class for ContextConnection
+* Add LocalContextConnection and LocalConnectionFactory
+* Remove connect(), disconnect() and getSocket() methods from Connection interface
+* Rename ContextConnection to RemoteContextConnection and ConnectionFactory to RemoteConnectionFactory
+
 # Version 0.6.1
 
 ## Bugfixes